requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.3
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
filelock==3.3.1
gunicorn==20.1.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.8.1
whitenoise==5.3.0
